:sparkles: feat: [회원가입] 로그인 Accesstoken 경로 변경 및 리펙터링
