+++ src/app/signup/page.tsx
@@ -1,69 +1,28 @@
 'use client';
-
 import * as React from 'react';
-import { SignupWrapper } from './style';
-import Image from 'next/image';
-import { css } from '@emotion/react';
+import { useEffect, useState } from 'react';
 
-import Button from '@/app/_component/atom/button/button';
-import { Colors, Icons, Images } from '@globalStyles';
-import { Fragment, useEffect, useState } from 'react';
-import { useRouter } from 'next/navigation';
+import { JoinWrap } from './style';
+
 import BackHeader from '@/app/_component/molecule/BackHeader';
-import InputForm from '@/app/_component/atom/InputForm';
-import FilterModal from '@/app/_component/organism/filterModal';
-import { agencyRanges, ageRanges, situationRanges } from '@/constants';
-import { OnChangeValueType, ParamsType } from '@/types/globalType';
-import { parseIdentity } from '@/hooks/useIdentity';
+import JoinTemplate from '@/app/_component/temp/JoinTemplate';
+import BottomButton from '@/app/_component/atom/BottomButton';
+import { OnChangeValueType } from '@/types/globalType';
+import * as queryString from 'querystring';
+import { useRouter } from 'next/navigation';
+import { useQueryParams } from '@/hooks/useParam';
+import { fetchAccessToken } from '@/hooks/useKakaoLogin';
 
-export default function Signup(): React.JSX.Element {
-  const [params, setParams] = useState<ParamsType>({
-    identity_first: '',
-    identity_last: '',
-    userName: '',
-    phoneNumber: '',
-    telecom: '',
-  });
-  // api 요청 시 identity_first 을 parseIdentity 사용하여 변환
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [openVarifi, setOpenVarifi] = useState(false);
+export default function Join(): React.JSX.Element {
   const router = useRouter();
-  console.log(params);
-  const onChangeValue: OnChangeValueType = (field, value) => {
-    setParams((prevState) => ({
-      ...prevState,
-      [field]: value,
-    }));
-  };
+  const { params, onChangeValue } = useQueryParams();
 
-  // param 모든 값이 존재하는지
-  const checkParamsFilled = () => {
-    const { identity_first, identity_last, userName, phoneNumber, telecom } =
-      params;
-    return (
-      identity_first && identity_last && userName && phoneNumber && telecom
-    );
-  };
-  const filterNumericInput = (
-    e: React.ChangeEvent<HTMLInputElement>,
-  ): string => {
-    return e.target.value.replace(/\D/g, ''); // 숫자가 아닌 문자를 필터링하여 제거
-  };
-
-  const handleNextButtonClick = () => {
-    if (checkParamsFilled()) {
-      setOpenVarifi(true);
-      router.push('/signup/verification');
+  const handleClick = () => {
+    if (params.signupState === true) {
+      router.push(`/signup/terms`);
+    } else {
+      router.push('/home');
     }
-  };
-
-  const handleAgencySelect = (selectedOptions: string[]) => {
-    onChangeValue('telecom', selectedOptions);
-    setIsModalOpen(false);
-  };
-  const resetAgencyOptions = () => {
-    onChangeValue('telecom', []);
   };
 
   const [code, setCode] = useState<string | null>(null);
@@ -77,142 +36,37 @@ export default function Signup(): React.JSX.Element {
 
   useEffect(() => {
     const fetchData = async () => {
-      try {
-        const response = await fetch(
-          `https://api-dev.vacgom.co.kr/api/v1/oauth/kakao/login?code=${code}`,
-        );
-        if (response.ok) {
-          const data = await response.json();
-          console.log(data);
-          localStorage.setItem('accessToken', data.token.accessToken);
-          setTimeout(() => {
-            router.push('/signup');
-          }, 1000);
-        } else {
-          console.error('Failed to fetch data:', response.status);
+      if (code) {
+        const response = await fetchAccessToken(code);
+        console.log(response);
+        if (response.accessToken) {
+          localStorage.setItem('accessToken', response.accessToken);
         }
-      } catch (error) {
-        console.error('Error fetching data:', error);
       }
     };
 
-    if (code) {
-      fetchData();
-    }
+    fetchData();
   }, [code]);
 
   return (
-    <SignupWrapper>
-      <BackHeader title={'회원가입'} url={'/vachistory'} />
-      <div className="top">정보를 입력해 주세요</div>
-      <div className="container">
-        <div className="item">
-          <InputForm
-            placeholder="이름"
-            value={params.userName}
-            descriptionTop={'이름'}
-            type="text"
-            onChange={(e) => {
-              onChangeValue('userName', filterNumericInput(e));
-            }}
-          />
-        </div>
-        <div className="item">
-          <InputForm
-            placeholder="통신사"
-            value={params.telecom}
-            descriptionTop={'통신사'}
-            rightIcon={Icons.arrow_down}
-            type="text"
-            customStyle={css`
-              & > .input__content > .input__content--right__icon > img {
-                width: 24px;
-                height: 24px;
-              }
-            `}
-            onClick={() => {
-              setIsModalOpen(true);
-            }}
-          />
-        </div>
-        <div className="item">
-          <InputForm
-            placeholder="번호 입력"
-            value={params.phoneNumber}
-            descriptionTop={'휴대폰 번호'}
-            type="text"
-            onChange={(e) => {
-              let filteredValue = filterNumericInput(e);
-              if (filteredValue.length > 11) {
-                filteredValue = filteredValue.slice(0, 11);
-              }
-              onChangeValue('phoneNumber', filteredValue);
-            }}
-          />
-        </div>
-        <div className="item">
-          <div className="input_title">주민등록번호</div>
-          <div className="item_row">
-            <InputForm
-              placeholder="YYMMDD"
-              value={params.identity_first}
-              type="text"
-              maxLength={6}
-              customStyle={css`
-                width: 50%;
-              `}
-              onChange={(e) => {
-                let filteredValue = filterNumericInput(e);
-                onChangeValue('identity_first', filteredValue);
-              }}
-            />
-            <p>-</p>
-            <InputForm
-              placeholder=""
-              value={params.identity_last}
-              type="text"
-              maxLength={1}
-              customStyle={css`
-                width: 60px;
-              `}
-              onChange={(e) => {
-                onChangeValue('identity_last', filterNumericInput(e));
-              }}
-            />
-            <div className="hiden_item">
-              <p></p>
-              <p></p>
-              <p></p>
-              <p></p>
-              <p></p>
-              <p></p>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <Fragment>
-        <FilterModal
-          isOpen={isModalOpen}
-          title="통신사를 선택해 주세요"
-          options={agencyRanges}
-          selectedOptions={params.telecom}
-          onClose={() => setIsModalOpen(false)}
-          onOptionSelect={handleAgencySelect}
-          onReset={resetAgencyOptions}
-        />
-      </Fragment>
-      <div className="bottom">
-        <button
-          className={
-            !checkParamsFilled() ? 'confirm_button' : 'confirm_button_Filled'
-          }
-          onClick={handleNextButtonClick}
-          disabled={!checkParamsFilled()}
-        >
-          다음
-        </button>
-      </div>
-    </SignupWrapper>
+    <JoinWrap>
+      <BackHeader title={' '} url={'/'} />
+      <JoinTemplate
+        title={'예방접종도우미에 가입한 적이 있나요?'}
+        subTop={'백곰을 이용하기 위해서는'}
+        subBottom={'질병관리청의 예방접종도우미 가입이 필요해요'}
+        falseLabel={'아니요, 가입한 적이 없어요'}
+        trueLabel={'네, 가입한 적이 있어요'}
+        params={params}
+        field={'signupState'}
+        onChangeValue={onChangeValue}
+      />
+      <BottomButton
+        filled={params.signupState === null}
+        handleNextButtonClick={() => {
+          handleClick();
+        }}
+      />
+    </JoinWrap>
   );
 }
