+++ src/app/signup/terms/page.tsx
@@ -17,12 +17,12 @@ export default function Terms(): React.JSX.Element {
   const { params, onChangeValue } = useQueryParams();
   console.log(params);
   const handleClick = () => {
-    router.push('/signup');
+    router.push('/signup/info');
   };
 
   return (
     <JoinWrap>
-      <BackHeader title={' '} url={'/signup/join'} />
+      <BackHeader title={' '} url={'/signup/info'} />
       <JoinTemplate
         title={'예방접종도우미의 \n' + '약관내용에 동의해주세요'}
         subTop={'회원약관에 동의해야 정상적으로 서비스를 이용할 수 있어요.'}
